Refactor Register implementation in preparation for RegisterStatic
